# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ILVOX Backend API",
    "version": "0.4.5",
    "description": "Contrato desplegado de Fase 4.5: organizaciones opcionales, conversión standalone y administración de servicios. No incluye contactos."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/leads": {
      "post": {
        "tags": ["Leads"],
        "summary": "Capture a public lead",
        "description": "Forces status=new and no assignee. Repeated email addresses are allowed. Limited to 10 requests/minute per application instance.",
        "security": [],
        "x-permission": "public",
        "x-scope": "public capture",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PublicLeadInput" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      },
      "get": {
        "tags": ["Leads"],
        "summary": "List authorized leads",
        "security": [{ "clerkSession": [] }],
        "x-permission": "leads.read",
        "x-scope": "global or assigned in SQL; internal actors only",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "$ref": "#/components/parameters/Search" },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/LeadStatus" } },
          { "name": "serviceId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "assignedToUserId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "createdFrom", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "createdTo", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          {
            "name": "sortBy",
            "in": "query",
            "schema": { "type": "string", "enum": ["createdAt", "updatedAt"], "default": "createdAt" }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/leads/{leadId}": {
      "get": {
        "tags": ["Leads"],
        "summary": "Get an authorized lead and redacted audit history",
        "security": [{ "clerkSession": [] }],
        "x-permission": "leads.read",
        "x-scope": "global or assigned in SQL; internal actors only",
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Leads"],
        "summary": "Update commercial lead fields",
        "description": "Status, assignment and conversion fields are forbidden.",
        "security": [{ "clerkSession": [] }],
        "x-permission": "leads.manage",
        "x-scope": "global or assigned in SQL; internal actors only",
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LeadCommercialPatch" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/leads/{leadId}/transition": {
      "post": {
        "tags": ["Leads"],
        "summary": "Apply a lead state-machine transition",
        "description": "converted is rejected here. Rejection and reopen require reason.",
        "security": [{ "clerkSession": [] }],
        "x-permission": "leads.manage",
        "x-scope": "global or assigned in SQL; internal actors only",
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "additionalProperties": false,
                "properties": {
                  "status": { "$ref": "#/components/schemas/LeadStatus" },
                  "reason": { "type": "string", "minLength": 1, "maxLength": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/leads/{leadId}/assign": {
      "post": {
        "tags": ["Leads"],
        "summary": "Assign an active internal user",
        "security": [{ "clerkSession": [] }],
        "x-permission": "leads.manage",
        "x-scope": "global or assigned in SQL; internal actors only",
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assignedToUserId"],
                "additionalProperties": false,
                "properties": {
                  "assignedToUserId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/leads/{leadId}/convert": {
      "post": {
        "tags": ["Leads"],
        "summary": "Convert an approved lead transactionally",
        "description": "Uses FOR UPDATE. standalone needs only leads.manage; organization modes also need organizations.manage. No mode creates a contact or Clerk identity.",
        "security": [{ "clerkSession": [] }],
        "x-permission": "leads.manage; plus organizations.manage for create_organization or reuse_organization",
        "x-scope": "global or assigned lead; organization scope only for organization modes",
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LeadConversionInput" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clerkSession": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Clerk session token"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "Search": {
        "name": "search",
        "in": "query",
        "schema": { "type": "string", "minLength": 1, "maxLength": 320 }
      },
...
      "LeadId": {
        "name": "leadId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "schemas": {
...
      "LeadSource": {
        "type": "string",
        "enum": ["diagnostic", "quotation", "contact", "referral", "campaign"]
      },
      "LeadStatus": {
        "type": "string",
        "enum": ["new", "contacted", "in_diagnostic", "quotation", "proposal_sent", "negotiation", "approved", "not_approved", "converted"]
      },
...
      "OrganizationSize": {
        "type": "string",
        "enum": ["micro", "small", "medium", "large"]
      },
      "PublicLeadInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["fullName", "email", "message", "source"],
        "properties": {
          "fullName": { "type": "string", "minLength": 1, "maxLength": 160 },
          "companyName": { "type": "string", "minLength": 1, "maxLength": 200 },
          "email": { "type": "string", "format": "email", "maxLength": 320 },
          "phone": { "type": "string", "minLength": 1, "maxLength": 40 },
          "serviceId": { "type": "string", "format": "uuid" },
          "message": { "type": "string", "minLength": 1, "maxLength": 5000 },
          "source": { "$ref": "#/components/schemas/LeadSource" }
        }
      },
      "LeadCommercialPatch": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "fullName": { "type": "string", "minLength": 1, "maxLength": 160 },
          "companyName": { "type": ["string", "null"], "maxLength": 200 },
          "email": { "type": "string", "format": "email", "maxLength": 320 },
          "phone": { "type": ["string", "null"], "maxLength": 40 },
          "serviceId": { "type": ["string", "null"], "format": "uuid" },
          "message": { "type": "string", "minLength": 1, "maxLength": 5000 },
          "source": { "$ref": "#/components/schemas/LeadSource" }
        }
      },
      "LeadConversionInput": {
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "required": ["mode"],
            "properties": {
              "mode": { "const": "standalone" }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": ["mode", "name"],
            "properties": {
              "mode": { "const": "create_organization" },
              "name": { "type": "string", "minLength": 1, "maxLength": 160 },
              "legalName": { "type": "string", "maxLength": 200 },
              "industry": { "type": "string", "maxLength": 120 },
              "size": { "$ref": "#/components/schemas/OrganizationSize" },
              "countryCode": { "type": "string", "pattern": "^[A-Z]{2}$" },
              "taxId": { "type": "string", "maxLength": 64 },
              "accountManagerUserId": { "type": "string", "format": "uuid" }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": ["mode", "organizationId"],
            "properties": {
              "mode": { "const": "reuse_organization" },
              "organizationId": { "type": "string", "format": "uuid" }
            }
          }
        ]
      },
...
      "ApiSuccess": {
        "type": "object",
        "required": ["data"],
        "properties": { "data": {} }
      },
      "ApiError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "requestId": { "type": "string" },
              "details": {}
            }
          }
        }
      }
    },
    "responses": {
      "Success": {
        "description": "Successful response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiSuccess" } } }
      },
      "ValidationError": {
        "description": "Validation failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
      "Unauthenticated": {
        "description": "Authentication required",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
      "Forbidden": {
        "description": "Permission or scope denied",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
      "NotFound": {
        "description": "Missing or outside authorized scope",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
      "Conflict": {
        "description": "Duplicate, invalid state, incompatible idempotent retry, or concurrency conflict",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
      "PayloadTooLarge": {
        "description": "Request body exceeds configured limit",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      }
    }
  }
}